feat(jdbc): Rename auth to connection
