+++ connectors/jdbc/src/main/java/io/camunda/connector/jdbc/model/request/connection/JdbcConnection.java
@@ -4,7 +4,7 @@
  * See the License.txt file for more information. You may not use this file
  * except in compliance with the proprietary license.
  */
-package io.camunda.connector.jdbc.model.request.auth;
+package io.camunda.connector.jdbc.model.request.connection;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
@@ -18,7 +18,7 @@ import io.camunda.connector.jdbc.model.request.SupportedDatabase;
 })
 @TemplateDiscriminatorProperty(
     label = "Connection type",
-    group = "authentication",
+    group = "connection",
     name = "authType",
     defaultValue = "uri")
 public sealed interface JdbcConnection permits UriConnection, DetailedConnection {
